New sprites imlemented

+++ BO360/BO360/BOPaddle.cpp
@@ -170,7 +170,7 @@ double BOPaddle::GetDegrees()const
 float2 BOPaddle::GetBallSpawnPosition()
 {
     float radius = (m_size.x * 0.5f);
-    float alpha = ((-m_rotation - (21 * (m_segments - 1))) * DEGREES_TO_RADIANS) + 2;
+    float alpha = (float)(((-m_rotation - (21 * (m_segments - 1))) * DEGREES_TO_RADIANS) + 2);
     float ballPosx = cos(alpha) * radius;
     float ballPosy = sin(alpha) * radius;
 
@@ -184,7 +184,7 @@ float2 BOPaddle::GetBallStuckPosition(float p_angle)
 {
     //Works
     float radius = (m_size.x * 0.5f);
-    float alpha = ((-m_rotation - p_angle) * DEGREES_TO_RADIANS) - 1.57;
+    float alpha = (float)(((-m_rotation - p_angle) * DEGREES_TO_RADIANS) - 1.57);
     float ballPosx = -cos(alpha) * radius;
     float ballPosy = sin(alpha) * radius;
     float tempx = m_position.x + ballPosx;
